export refactoring

+++ lib/features/book/presentation/book_page_controller.cg.dart
@@ -1,10 +1,11 @@
 import 'dart:io';
 import 'dart:typed_data';
 
-import 'package:file_saver/file_saver.dart';
+import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:mobx/mobx.dart';
 import 'package:path_provider/path_provider.dart';
+import 'package:permission_handler/permission_handler.dart';
 import 'package:share_plus/share_plus.dart';
 
 import '../../../core/di.dart';
@@ -83,9 +84,9 @@ abstract class BookPageControllerBase with Store {
 
       bookXmlBytes = await convertToFB2(
         book.value!,
-        (prgrs) {
-          progress = prgrs;
-          logger.i('Progress: $prgrs');
+        (progress) {
+          this.progress = progress;
+          logger.i('Progress: $progress');
         },
       );
     } catch (e, trace) {
@@ -107,52 +108,54 @@ abstract class BookPageControllerBase with Store {
     name = name.replaceAll('?', '');
     name = name.replaceAll(':', '');
 
-    if (share) {
-      var tempDir = await getApplicationSupportDirectory();
-      final path = '${tempDir.path}/$name.fb2';
-      var file = File(path);
-      await file.writeAsBytes(bookXmlBytes!);
+    var tempDir = (await getTemporaryDirectory()).path;
+    final path = '$tempDir\\$name.fb2';
+    var file = File(path);
+    await file.writeAsBytes(bookXmlBytes!);
 
-      try {
-        var xfile = XFile(
-          path,
-          mimeType: 'application/x-fictionbook+xml',
-        );
+    try {
+      final xfile = XFile(
+        path,
+        name: '$name.fb2',
+        mimeType: 'application/x-fictionbook+xml',
+      );
 
+      if (share) {
         final text = '${data.title}'
             '\nАвторы: ${data.authors.map((e) => e.name).join(', ')}'
             '${data.series == null ? '' : '\nСерия: ${data.series!.name} #${data.series!.number}'}'
             '\n'
-            '\n${data.chapters.length} глав'
-            '\nПо: "${data.chapters.last.title}"';
+            '\nГлав: ${data.chapters.length}';
 
         await Share.shareXFiles([xfile], text: text, subject: text);
-      } finally {
-        await file.delete();
+        return;
       }
-    } else {
-      final bool saveAs =
-          Platform.isAndroid || Platform.isIOS || Platform.isMacOS;
-      final func =
-          saveAs ? FileSaver.instance.saveAs : FileSaver.instance.saveFile;
+      bool isGranted =
+          await Permission.manageExternalStorage.request().isGranted;
+      if (!isGranted) openAppSettings();
 
-      try {
-        await func(
-          name: name,
-          ext: 'fb2',
-          bytes: bookXmlBytes!,
-          mimeType: MimeType.custom,
-          customMimeType: 'application/x-fictionbook+xml',
+      final finalPath = await FilePicker.platform.saveFile(
+        dialogTitle: 'Сохранение книги',
+        fileName: '$name.fb2',
+        bytes: bookXmlBytes,
+        type: FileType.custom,
+        allowedExtensions: ['.fb2'],
+      );
+
+      if (finalPath == null) {
+        return overlaySnackMessage(
+          scaffoldKey.currentContext!,
+          'Сохранение отменено',
         );
-        if (scaffoldKey.currentContext != null && !saveAs) {
-          overlaySnackMessage(
-            scaffoldKey.currentContext!,
-            'Книга сохранена в загрузки',
-          );
-        }
-      } catch (e, trace) {
-        logger.e('Error saving book', error: e, stackTrace: trace);
       }
+
+      await xfile.saveTo(finalPath);
+      overlaySnackMessage(scaffoldKey.currentContext!, 'Успешно сохранено');
+    } catch (e, trace) {
+      overlaySnackMessage(scaffoldKey.currentContext!, 'Произошла ошибка');
+      logger.e('Error saving book', error: e, stackTrace: trace);
+    } finally {
+      if (!Platform.isWindows) await file.delete();
     }
   }
 }
